Django/requirements.txt
asgiref==3.4.1
attrs==21.4.0
cached-property==1.5.2
click==8.0.3
colorama==0.4.4
Django==4.0.1
djangorestframework==3.13.1
djangorestframework-camel-case==1.3.0
djangorestframework-simplejwt==5.0.0
drf-spectacular==0.21.1
graphql-core==3.1.7
inflection==0.5.1
jsonschema==4.4.0
Pygments==2.11.2
PyJWT==2.3.0
pyrsistent==0.18.1
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2021.3
PyYAML==6.0
sentinel==0.3.0
six==1.16.0
sqlparse==0.4.2
strawberry-graphql==0.93.23
strawberry-graphql-django==0.2.5
typing_extensions==4.0.1
tzdata==2021.5
uritemplate==4.1.1
